requirements.txt
python-telegram-bot==20.0a4
requests==2.28.1
marshmallow-dataclass==8.5.9
